refactor: modularize main.rs into organized CLI structure
Breaking down the monolithic 1,450-line main.rs into maintainable modules:

- Extract CLI argument definitions to cli/args.rs
- Move validation logic to cli/validation.rs
- Separate configuration loading to cli/config.rs
- Create command modules: analyze.rs, batch.rs, benchmark.rs, database.rs
- Move error handling to error/handler.rs
- Reorganize output modules: display.rs, batch_results.rs
- Reduce main.rs from 1,450 → 42 lines (97% reduction)

All tests pass, functionality preserved. First of 3 planned refactors.
