+++ src/analysis/mod.rs
@@ -6,7 +6,7 @@ pub mod patterns;
 pub use column::{analyze_column, analyze_column_fast};
 pub use inference::infer_type;
 pub use ml_readiness::{
-    FeatureAnalysis, MlReadinessEngine, MlReadinessLevel, MlReadinessScore, MlRecommendation,
-    PreprocessingSuggestion,
+    FeatureAnalysis, MlBlockingIssue, MlReadinessEngine, MlReadinessLevel, MlReadinessScore,
+    MlRecommendation, PreprocessingSuggestion, RecommendationPriority,
 };
 pub use patterns::detect_patterns;
