+++ src/commands/database.rs
@@ -1,0 +1,98 @@
+use anyhow::Result;
+use colored::*;
+
+use crate::cli::args::Cli;
+use crate::output::{display_profile, display_quality_issues};
+use dataprof::{generate_html_report, profile_database, DatabaseConfig};
+
+#[cfg(feature = "database")]
+pub fn run_database_analysis(cli: &Cli, connection_string: &str) -> Result<()> {
+    use tokio;
+
+    println!(
+        "{}",
+        "🗃️ DataProfiler v0.3.0 - Database Analysis"
+            .bright_blue()
+            .bold()
+    );
+    println!();
+
+    // Default query or table (use file parameter as table name if no query provided)
+    let query = if let Some(sql_query) = cli.query.as_ref() {
+        sql_query.to_string()
+    } else {
+        let table_name = cli.file.display().to_string();
+        if table_name.is_empty() || table_name == "." {
+            return Err(anyhow::anyhow!("Please specify either --query 'SELECT * FROM table' or provide table name as file argument"));
+        }
+        format!("SELECT * FROM {}", table_name)
+    };
+
+    // Create database configuration
+    let config = DatabaseConfig {
+        connection_string: connection_string.to_string(),
+        batch_size: cli.batch_size,
+        max_connections: Some(10),
+        connection_timeout: Some(std::time::Duration::from_secs(30)),
+        retry_config: Some(dataprof::database::RetryConfig::default()),
+        sampling_config: None,
+        enable_ml_readiness: true,
+        ssl_config: Some(dataprof::database::SslConfig::default()),
+        load_credentials_from_env: true,
+    };
+
+    // Run async analysis
+    let rt = tokio::runtime::Runtime::new()
+        .map_err(|e| anyhow::anyhow!("Failed to create async runtime: {}", e))?;
+
+    let report = rt.block_on(async { profile_database(config, &query).await })?;
+
+    println!(
+        "🔗 {} | {} columns | {} rows",
+        connection_string
+            .split('@')
+            .next_back()
+            .unwrap_or(connection_string),
+        report.file_info.total_columns,
+        report.file_info.total_rows.unwrap_or(0)
+    );
+
+    if report.scan_info.rows_scanned > 0 {
+        let scan_time_sec = report.scan_info.scan_time_ms as f64 / 1000.0;
+        let rows_per_sec = report.scan_info.rows_scanned as f64 / scan_time_sec;
+        println!(
+            "⚡ Processed {} rows in {:.1}s ({:.0} rows/sec)",
+            report.scan_info.rows_scanned, scan_time_sec, rows_per_sec
+        );
+    }
+    println!();
+
+    if cli.quality {
+        // Show quality issues
+        display_quality_issues(&report.issues);
+
+        // Generate HTML report if requested
+        if let Some(html_path) = &cli.html {
+            match generate_html_report(&report, html_path) {
+                Ok(_) => {
+                    println!(
+                        "📄 HTML report saved to: {}",
+                        html_path.display().to_string().bright_green()
+                    );
+                    println!();
+                }
+                Err(e) => {
+                    eprintln!("❌ Failed to generate HTML report: {}", e);
+                }
+            }
+        }
+    }
+
+    // Show column profiles
+    for profile in &report.column_profiles {
+        display_profile(profile);
+        println!();
+    }
+
+    Ok(())
+}
