+++ src/commands/mod.rs
@@ -1,0 +1,13 @@
+pub mod analyze;
+pub mod batch;
+pub mod benchmark;
+
+#[cfg(feature = "database")]
+pub mod database;
+
+pub use analyze::{is_json_file, run_analysis};
+pub use batch::{create_batch_config, run_batch_directory, run_batch_glob};
+pub use benchmark::{run_benchmark_analysis, show_engine_info};
+
+#[cfg(feature = "database")]
+pub use database::run_database_analysis;
