+++ src/error/mod.rs
@@ -1,0 +1,3 @@
+pub mod handler;
+
+pub use handler::*;
