+++ src/lib.rs
@@ -44,7 +44,8 @@ pub use engines::{AdaptiveProfiler, EnginePerformance, ProcessingType};
 // Re-exports for backward compatibility
 pub use output::html::generate_html_report;
 pub use types::{
-    ColumnProfile, ColumnStats, DataType, FileInfo, Pattern, QualityIssue, QualityReport, ScanInfo,
+    ColumnProfile, ColumnStats, DataType, FileInfo, OutputFormat, Pattern, QualityIssue,
+    QualityReport, ScanInfo, Severity,
 };
 pub use utils::quality::QualityChecker;
 pub use utils::sampler::{SampleInfo, Sampler};
