+++ src/output/mod.rs
@@ -1,3 +1,13 @@
+pub mod batch_results;
+pub mod display;
 pub mod formatters;
 pub mod html;
 pub mod progress;
+
+pub use batch_results::display_batch_results;
+pub use display::{
+    display_ml_score, display_profile, display_quality_issues, output_json_profiles,
+};
+pub use formatters::output_with_formatter;
+pub use html::*;
+pub use progress::*;
